fix: MFE service unavailable behind custom host proxy
The MFE service was running directly behind Caddy, and not Nginx. Thus, when
Caddy was disabled (RUN_CADDY=false), the MFE service was unavailable. To
address this, we include nginx as an extra proxy between Caddy and the MFE.
Caddy should only take care of SSL termination, as for other services.

Close #6.