feat: Add support for the Course Authoring MFE
The only feature enabled by default is the "Pages & Resources" menu item
in Studio.

Waffle flags for the new React editors are created but disabled by
default.  (The only editor that seems to be fully featured at the moment
is the HTML block editor; Video seems to work without a preview, and
Problem is just a placeholder.)

The new proctoring interface is also disabled, as it requires the
`edx-exams` IDA, which Tutor doesn't currently support.