fix: lower MFE config API cache timeout
Calls to the MFE config API endpoint are made with two caches: one on
the client side, in the "edx-cache" local storage, and one on the server
side. The client side cache has a hard-coded timeout of 5 minutes. I see
no reason to have such a high cache timeout on the server side.