fix: pngquant error during build on linux/arm64
Building the mfe Docker image on arm64 triggers the following error in
the 'account-prod' stage:

    Cannot find module 'imagemin-pngquant'

This was due to a missing system package.

I find it strange that no one reported this issue before, when there are
plenty of users running macOS with arm64. I discovered this issue while
setting up a multi-arch build CI environment.