fix: remove trailing slash from PUBLIC_PATH env var
We previously generated the PUBLIC_PATH base for each MFE with a
trailing slash, e.g. "/authn/". But the actual links that point to
these from the LMS and Studio don't have a trailing slash in them,
e.g. "/authn". This commit removes the trailing slash to be consistent.
Removing that trailing slash in configuration also means that we don't
have to strip it out on the JavaScript side of things, which would have
been necessary after the React Router 6 upgrade (we send it as the
router's basename).