fix: Runtime config in dev mode
MFE runtime configuration was not working properly prior to this.  It
would always end up querying the `example.com` site configuration.  With
this change, we force the request origin to always be the LMS host and
port, as opposed to something like `apps.local.edly.io:2001`, which
would trip up the Django site detection.